requirements.txt
fastapi==0.115.2
python-dotenv==1.1.0
uvicorn==0.34.0
loguru==0.7.3
numpy==2.3.1
pandas==2.3.0
plotly==6.1.2
